requirements.txt
# Core ML/AI
torch>=2.0.0
transformers>=4.35.0
numpy>=1.24.0

# Audio Processing
librosa>=0.10.0
soundfile>=0.12.0
scipy>=1.10.0

# Video Processing
yt-dlp>=2023.11.16
opencv-python>=4.8.0
ffmpeg-python>=0.2.0

# Computer Vision (for Visual-Audio Isolation)
ultralytics>=8.0.0

# Speech Processing (for Universal Broadcaster)
openai-whisper>=20231117

# Utilities
tqdm>=4.65.0
python-dotenv>=1.0.0
